requirements.txt
wtforms
flask
netifaces
genpy
python-nmap==0.6.1
resource